WIP: Support multiple schema properties
The last remaining task is to fix `setValue` (if needed)
