+++ app/ide-desktop/lib/dashboard/src/components/JSONSchemaInput.tsx
@@ -1,0 +1,401 @@
+/** @file A dynamic wizard for creating an arbitrary type of Data Link. */
+import * as React from 'react'
+
+import * as backendProvider from '#/providers/BackendProvider'
+
+import Autocomplete from '#/components/Autocomplete'
+import Dropdown from '#/components/Dropdown'
+
+import * as jsonSchema from '#/utilities/jsonSchema'
+import * as object from '#/utilities/object'
+
+// =====================
+// === getSchemaName ===
+// =====================
+
+const SCHEMA_NAMES = new WeakMap<object, string>()
+
+/** Return a human-readable name representing a schema. */
+function getSchemaNameHelper(defs: Record<string, object>, schema: object): string {
+  if ('title' in schema) {
+    return String(schema.title)
+  } else if ('type' in schema) {
+    return String(schema.type)
+  } else if ('$ref' in schema) {
+    const referencedSchema = jsonSchema.lookupDef(defs, schema)
+    return referencedSchema == null ? '(unknown)' : getSchemaName(defs, referencedSchema)
+  } else if ('anyOf' in schema) {
+    const members = Array.isArray(schema.anyOf) ? schema.anyOf : []
+    return (
+      members
+        .flatMap(object.singletonObjectOrNull)
+        .map(childSchema => getSchemaName(defs, childSchema))
+        .join(' | ') || '(unknown)'
+    )
+  } else if ('allOf' in schema) {
+    const members = Array.isArray(schema.allOf) ? schema.allOf : []
+    return (
+      members
+        .flatMap(object.singletonObjectOrNull)
+        .map(childSchema => getSchemaName(defs, childSchema))
+        .join(' & ') || '(unknown)'
+    )
+  } else {
+    return '(unknown)'
+  }
+}
+
+/** Return a human-readable name representing a schema.
+ * This function is a memoized version of {@link getSchemaNameHelper}. */
+function getSchemaName(defs: Record<string, object>, schema: object) {
+  const cached = SCHEMA_NAMES.get(schema)
+  if (cached != null) {
+    return cached
+  } else {
+    const name = getSchemaNameHelper(defs, schema)
+    SCHEMA_NAMES.set(schema, name)
+    return name
+  }
+}
+
+// =======================
+// === JSONSchemaInput ===
+// =======================
+
+/** Props for a {@link JSONSchemaInput}. */
+export interface JSONSchemaInputProps {
+  readonly dropdownTitle?: string
+  readonly defs: Record<string, object>
+  readonly readOnly?: boolean
+  readonly schema: object
+  readonly path: string
+  readonly getValidator: (path: string) => (value: unknown) => boolean
+  readonly value: NonNullable<unknown> | null
+  readonly setValue: React.Dispatch<React.SetStateAction<NonNullable<unknown> | null>>
+}
+
+/** A dynamic wizard for creating an arbitrary type of Data Link. */
+export default function JSONSchemaInput(props: JSONSchemaInputProps) {
+  const { dropdownTitle, readOnly = false, defs, schema, path, getValidator } = props
+  const { value: valueRaw, setValue: setValueRaw } = props
+  // The functionality for inputting `enso-secret`s SHOULD be injected using a plugin,
+  // but it is more convenient to avoid having plugin infrastructure.
+  const { backend } = backendProvider.useBackend()
+  const [value, setValue] = React.useState(valueRaw)
+  const [autocompleteText, setAutocompleteText] = React.useState(() =>
+    typeof value === 'string' ? value : null
+  )
+  const [selectedChildIndex, setSelectedChildIndex] = React.useState<number | null>(null)
+  const [autocompleteItems, setAutocompleteItems] = React.useState<string[] | null>(null)
+
+  React.useEffect(() => {
+    setValue(valueRaw)
+    // `initializing` is not a dependency.
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [valueRaw])
+
+  React.useEffect(() => {
+    setValueRaw(value)
+    // `setStateRaw` is a callback, not a dependency.
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [value])
+
+  // NOTE: `enum` schemas omitted for now as they are not yet used.
+  if ('const' in schema) {
+    // This value cannot change.
+    return null
+  } else {
+    const children: JSX.Element[] = []
+    if ('type' in schema) {
+      switch (schema.type) {
+        case 'string': {
+          if ('format' in schema && schema.format === 'enso-secret') {
+            const isValid = typeof value === 'string' && value !== ''
+            if (autocompleteItems == null) {
+              setAutocompleteItems([])
+              void (async () => {
+                const secrets = await backend.listSecrets()
+                // FIXME: Extract secret path instead of ID.
+                setAutocompleteItems(secrets.map(secret => secret.id))
+              })()
+            }
+            children.push(
+              <div
+                className={`rounded-2xl border ${
+                  isValid ? 'border-black/10' : 'border-red-700/60'
+                }`}
+              >
+                <Autocomplete
+                  items={autocompleteItems ?? []}
+                  itemToKey={item => item}
+                  itemToString={item => item}
+                  placeholder="Enter secret path"
+                  matches={(item, text) => item.toLowerCase().includes(text.toLowerCase())}
+                  values={isValid ? [value] : []}
+                  setValues={values => {
+                    setValue(values[0])
+                  }}
+                  text={autocompleteText}
+                  setText={setAutocompleteText}
+                />
+              </div>
+            )
+          } else {
+            children.push(
+              <input
+                type="text"
+                readOnly={readOnly}
+                value={typeof value === 'string' ? value : ''}
+                size={1}
+                className={`rounded-full w-40 px-2 bg-transparent border leading-170 h-6 py-px disabled:opacity-50 read-only:opacity-75 read-only:cursor-not-allowed ${
+                  getValidator(path)(value) ? 'border-black/10' : 'border-red-700/60'
+                }`}
+                placeholder="Enter text"
+                onChange={event => {
+                  const newValue: string = event.currentTarget.value
+                  setValue(newValue)
+                }}
+              />
+            )
+          }
+          break
+        }
+        case 'number': {
+          children.push(
+            <input
+              type="number"
+              readOnly={readOnly}
+              value={typeof value === 'number' ? value : ''}
+              size={1}
+              className={`rounded-full w-40 px-2 bg-transparent border leading-170 h-6 py-px disabled:opacity-50 read-only:opacity-75 read-only:cursor-not-allowed ${
+                getValidator(path)(value) ? 'border-black/10' : 'border-red-700/60'
+              }`}
+              placeholder="Enter number"
+              onChange={event => {
+                const newValue: number = event.currentTarget.valueAsNumber
+                if (Number.isFinite(newValue)) {
+                  setValue(newValue)
+                }
+              }}
+            />
+          )
+          break
+        }
+        case 'integer': {
+          children.push(
+            <input
+              type="number"
+              readOnly={readOnly}
+              value={typeof value === 'number' ? value : ''}
+              size={1}
+              className={`rounded-full w-40 px-2 bg-transparent border leading-170 h-6 py-px disabled:opacity-50 read-only:opacity-75 read-only:cursor-not-allowed ${
+                getValidator(path)(value) ? 'border-black/10' : 'border-red-700/60'
+              }`}
+              placeholder="Enter integer"
+              onChange={event => {
+                const newValue: number = Math.floor(event.currentTarget.valueAsNumber)
+                if (Number.isFinite(newValue)) {
+                  setValue(newValue)
+                }
+              }}
+            />
+          )
+          break
+        }
+        case 'boolean': {
+          children.push(
+            <input
+              type="checkbox"
+              readOnly={readOnly}
+              checked={typeof value === 'boolean' && value}
+              onChange={event => {
+                setValue(event.currentTarget.checked)
+              }}
+            />
+          )
+          break
+        }
+        case 'object': {
+          const propertiesObject =
+            'properties' in schema ? object.asObject(schema.properties) ?? {} : {}
+          const requiredProperties =
+            'required' in schema && Array.isArray(schema.required) ? schema.required : []
+          const propertyDefinitions = Object.entries(propertiesObject).flatMap(
+            (kv: [string, unknown]) => {
+              const [k, v] = kv
+              return object
+                .singletonObjectOrNull(v)
+                .map(childSchema => ({ key: k, schema: childSchema }))
+            }
+          )
+          if (jsonSchema.constantValue(defs, schema).length !== 1) {
+            children.push(
+              <div className="flex flex-col gap-1 rounded-2xl border border-black/10 p-2">
+                {propertyDefinitions.map(definition => {
+                  const { key, schema: childSchema } = definition
+                  const isOptional = !requiredProperties.includes(key)
+                  return jsonSchema.constantValue(defs, childSchema).length === 1 ? null : (
+                    <div
+                      key={key}
+                      className="flex flex-wrap items-center"
+                      {...('description' in childSchema
+                        ? { title: String(childSchema.description) }
+                        : {})}
+                    >
+                      <div
+                        className={`inline-block h-6 leading-170 py-px w-28 whitespace-nowrap ${
+                          isOptional ? 'cursor-pointer' : ''
+                        } ${value != null && key in value ? '' : 'opacity-50'}`}
+                        onClick={() => {
+                          if (isOptional) {
+                            setValue(oldValue => {
+                              if (oldValue != null && key in oldValue) {
+                                // This is SAFE, as `value` is an untyped object.
+                                // The removed key is intentionally unused.
+                                // eslint-disable-next-line no-restricted-syntax, @typescript-eslint/no-unused-vars
+                                const { [key]: removed, ...newValue } = oldValue as Record<
+                                  string,
+                                  NonNullable<unknown> | null
+                                >
+                                return newValue
+                              } else {
+                                return {
+                                  ...oldValue,
+                                  [key]: jsonSchema.constantValue(defs, childSchema, true)[0],
+                                }
+                              }
+                            })
+                          }
+                        }}
+                      >
+                        {'title' in childSchema ? String(childSchema.title) : key}
+                      </div>
+                      {value != null && key in value && (
+                        <JSONSchemaInput
+                          readOnly={readOnly}
+                          defs={defs}
+                          schema={childSchema}
+                          path={`${path}/properties/${key}`}
+                          getValidator={getValidator}
+                          // This is SAFE, as `value` is an untyped object.
+                          // eslint-disable-next-line no-restricted-syntax
+                          value={(value as Record<string, unknown>)[key] ?? null}
+                          setValue={newValue => {
+                            setValue(oldValue =>
+                              typeof oldValue === 'object' &&
+                              oldValue != null &&
+                              // This is SAFE; but there is no way to tell TypeScript that an object
+                              // has an index signature.
+                              // eslint-disable-next-line no-restricted-syntax
+                              (oldValue as Record<string, unknown>)[key] === newValue
+                                ? oldValue
+                                : { ...oldValue, [key]: newValue }
+                            )
+                          }}
+                        />
+                      )}
+                    </div>
+                  )
+                })}
+              </div>
+            )
+          }
+          break
+        }
+      }
+    }
+    if ('$ref' in schema && typeof schema.$ref === 'string') {
+      const referencedSchema = jsonSchema.lookupDef(defs, schema)
+      if (referencedSchema != null) {
+        children.push(
+          <JSONSchemaInput
+            {...props}
+            key={schema.$ref}
+            schema={referencedSchema}
+            path={schema.$ref}
+          />
+        )
+      }
+    }
+    if ('anyOf' in schema && Array.isArray(schema.anyOf)) {
+      const childSchemas = schema.anyOf.flatMap(object.singletonObjectOrNull)
+      const selectedChildSchema =
+        selectedChildIndex == null ? null : childSchemas[selectedChildIndex]
+      const selectedChildPath = `${path}/anyOf/${selectedChildIndex}`
+      const childValue =
+        selectedChildSchema == null ? [] : jsonSchema.constantValue(defs, selectedChildSchema)
+      if (
+        value != null &&
+        (selectedChildSchema == null || getValidator(selectedChildPath)(value) !== true)
+      ) {
+        const newIndexRaw = childSchemas.findIndex((_, index) =>
+          getValidator(`${path}/anyOf/${index}`)(value)
+        )
+        const newIndex = selectedChildSchema == null && newIndexRaw === -1 ? 0 : newIndexRaw
+        if (newIndex !== -1 && newIndex !== selectedChildIndex) {
+          setSelectedChildIndex(newIndex)
+        }
+      }
+      const dropdown = (
+        <Dropdown
+          readOnly={readOnly}
+          items={childSchemas}
+          selectedIndex={selectedChildIndex}
+          render={childProps => getSchemaName(defs, childProps.item)}
+          className="self-start"
+          onClick={(childSchema, index) => {
+            setSelectedChildIndex(index)
+            const newConstantValue = jsonSchema.constantValue(defs, childSchema, true)
+            setValue(newConstantValue[0] ?? null)
+            setSelectedChildIndex(index)
+          }}
+        />
+      )
+      children.push(
+        <div className={`flex flex-col gap-1 ${childValue.length === 0 ? 'w-full' : ''}`}>
+          {dropdownTitle != null ? (
+            <div className="flex items-center">
+              <div className="w-12 h-6 py-1">{dropdownTitle}</div>
+              {dropdown}
+            </div>
+          ) : (
+            dropdown
+          )}
+          {selectedChildSchema != null && (
+            <JSONSchemaInput
+              key={selectedChildIndex}
+              defs={defs}
+              readOnly={readOnly}
+              schema={selectedChildSchema}
+              path={selectedChildPath}
+              getValidator={getValidator}
+              value={value}
+              setValue={setValue}
+            />
+          )}
+        </div>
+      )
+    }
+    if ('allOf' in schema && Array.isArray(schema.allOf)) {
+      const childSchemas = schema.allOf.flatMap(object.singletonObjectOrNull)
+      const newChildren = childSchemas.map((childSchema, i) => (
+        <JSONSchemaInput
+          key={i}
+          defs={defs}
+          readOnly={readOnly}
+          schema={childSchema}
+          path={`${path}/allOf/${i}`}
+          getValidator={getValidator}
+          value={value}
+          setValue={setValue}
+        />
+      ))
+      children.push(...newChildren)
+    }
+    return children.length === 1 && children[0] != null ? (
+      children[0]
+    ) : (
+      <div className="flex flex-col gap-1">{...children}</div>
+    )
+  }
+}
