+++ app/ide-desktop/lib/dashboard/src/layouts/dashboard/UpsertDataLinkModal.tsx
@@ -5,8 +5,7 @@ import SCHEMA from '#/data/dataLinkSchema.json' assert { type: 'json' }
 
 import * as modalProvider from '#/providers/ModalProvider'
 
-import DataLinkInput from '#/layouts/dashboard/DataLinkInput'
-
+import DataLinkInput from '#/components/dashboard/DataLinkInput'
 import Modal from '#/components/Modal'
 
 import * as jsonSchema from '#/utilities/jsonSchema'
